tools/riposte-cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "riposte-cli"
version = "1.0.0"
description = "CLI tool for annotating meme images with GitHub Copilot AI for Riposte Android app"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Riposte", email = "riposte@example.com" }
]
keywords = ["meme", "emoji", "ai", "copilot", "image-annotation", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "github-copilot-sdk>=0.1.18",
    "Pillow>=10.0.0",
    "imagehash>=4.3.0",
    "filetype>=1.2.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/riposte/riposte"
Repository = "https://github.com/riposte/riposte"
Issues = "https://github.com/riposte/riposte/issues"

[project.scripts]
meme-cli = "riposte_cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/riposte_cli"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
